% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_process_subfuncs.R
\name{combine_study_facts}
\alias{combine_study_facts}
\title{combine PF output into one table per study}
\usage{
combine_study_facts(
  pf_tbl,
  study_abbr,
  domain_list,
  time = FALSE,
  visit_type_list = list("inpatient", "outpatient", "other_visit", "all")
)
}
\arguments{
\item{pf_tbl}{the table output by \code{compute_pf} or \code{compute_pf_for_fot}}

\item{study_abbr}{string that corresponds with the study/studies of interest}

\item{domain_list}{a CSV file with configuration information about each
domain of interest}

\item{time}{a logical indicating whether the analysis was conducted over time}

\item{visit_type_list}{the types of visits in the PF output; defaults to:
inpatient, outpatient, other_visit, all}
}
\value{
a list of dataframes that contain all sites and all domains for each study
listed in \code{study_abbr}. the resulting dataframes will also have an
age grouping column added.
}
\description{
combine PF output into one table per study
}
