% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_process_subfuncs.R
\name{compute_dist_mean_pf}
\alias{compute_dist_mean_pf}
\title{Compute distance from mean}
\usage{
compute_dist_mean_pf(
  data_input,
  site_col,
  n_sd = 2,
  agegrp = NULL,
  codeset = NULL
)
}
\arguments{
\item{data_input}{input table generated by \code{loop_through_visits} and reduced with \code{combine_study_facts}}

\item{site_col}{the column in the data that holds site names}

\item{n_sd}{the number of standard deviations that should be used as a threshold to
identify an outlier}

\item{agegrp}{a csv file with user designated age groupings, based on age at cohort entry}

\item{codeset}{a csv file with user designated cohort flags based on a user provided codeset}
}
\value{
a dataframe that summarises the number of patients with fact counts that fall +/- 3 SD away from the mean
both at fact and site + fact levels

\if{html}{\out{<div class="sourceCode">}}\preformatted{    contains columns: person_id, start_date, end_date, fu, site, domain, var_val, var_ever, study,
                      visit_type, n_fact, zscore_fact, outlier_fact, prop_outlier_fact, n_site_fact,
                      zscore_site_fact, outlier_site_fact, prop_outlier_site_fact
}\if{html}{\out{</div>}}
}
\description{
Compute distance from mean
}
