% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_process_subfuncs.R
\name{compute_pf_medians}
\alias{compute_pf_medians}
\title{Compute median fact counts by visit & fact type and by site, visit, & fact type}
\usage{
compute_pf_medians(data_input, site_col, agegrp = NULL, codeset = NULL)
}
\arguments{
\item{data_input}{the list output of \code{combine_study_facts}}

\item{site_col}{the column in the data that holds site names}

\item{agegrp}{boolean that determines whether the output should also
be grouped by age group}

\item{codeset}{boolean that determines whether the output should also
be grouped by a user provided codeset flag}
}
\value{
dataframe that contains the total median number of facts for that visit & fact_type
as well as the median number of facts for the site, visit, & fact type for each
study
}
\description{
Compute median fact counts by visit & fact type and by site, visit, & fact type
}
