% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_process_subfuncs_omop.R
\name{compute_pf_omop}
\alias{compute_pf_omop}
\title{Compute patient facts per year of cohort follow up for each domain
defined in the domain_tbl}
\usage{
compute_pf_omop(cohort, pf_input_tbl, grouped_list, domain_tbl)
}
\arguments{
\item{cohort}{the cohort of patients of interest for the study}

\item{pf_input_tbl}{a table with the relevant domain of interest
filtered to the desired visit type of interest}

\item{grouped_list}{a vector of column names by which the input should be grouped}

\item{domain_tbl}{a table with definitions for each domain (i.e. patientfacts::pf_domain_file)}
}
\value{
a dataframe with one column for each of the specified domains associated with
the visit type of interest summarizing the facts per year of follow up for each
patient in the cohort
}
\description{
Compute patient facts per year of cohort follow up for each domain
defined in the domain_tbl
}
\keyword{internal}
