% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_process_subfuncs_pcornet.R
\name{compute_pf_pcnt}
\alias{compute_pf_pcnt}
\title{function to take in domain and compute patient facts}
\usage{
compute_pf_pcnt(cohort, pf_input_tbl, grouped_list, domain_tbl)
}
\arguments{
\item{cohort}{the cohort for which to iterate}

\item{pf_input_tbl}{the tbl that should be iterated
so person facts are computed}

\item{grouped_list}{a vector to group input by. Defaults to \code{person_id},
\code{start_date}, \code{end_date}, \code{fu}, \code{site}}

\item{domain_tbl}{the config CSV file}
}
\value{
the column \code{person_id}, for the specified visit type,
the number of facts for the person - each fact corresponds to the
fact in the CSV file; each column is a domain
}
\description{
function to take in domain and compute patient facts
}
