% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pf_domain_file}
\alias{pf_domain_file}
\title{Patient Facts Domain File Sample}
\format{
\subsection{pf_domain_file}{

A dataframe with 3 columns:
\describe{
\item{domain}{An arbitrary string to label the domain of interest}
\item{domain_tbl}{The name of the CDM table associated with the domain of interest}
\item{filter_logic}{(optional) a string to be parsed as logic to filter the domain_tbl as needed to best represent the domain}
}
}
}
\usage{
pf_domain_file
}
\description{
A sample version of the file structure expected for the \code{domain_tbl}
parameter in the \code{pf_process} function. The user should recreate
this file and include their own domain definitions.
}
\keyword{datasets}
