% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output_subfuncs.R
\name{pf_ms_anom_at}
\alias{pf_ms_anom_at}
\title{\strong{Multi-Site, Anomaly, Over Time}}
\usage{
pf_ms_anom_at(process_output, domain_filter, visit_filter)
}
\arguments{
\item{process_output}{output of \code{pf_process} function}

\item{domain_filter}{one of the user provided domains in the process_output table to be used
to filter down the output}

\item{visit_filter}{one of the user provided visit types in the process_output table to be used
to filter down the output}
}
\value{
three graphs:
\enumerate{
\item line graph that shows the smoothed proportion of a
domain across time computation with the Euclidean distance associated with each line
\item line graph that shows the raw proportion of a
domain across time computation with the Euclidean distance associated with each line
\item a bar graph with the Euclidean distance value for each site, with the average
proportion as the fill
}
}
\description{
Proportion of patients with fact & Euclidean distance summary
}
