% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output_subfuncs.R
\name{pf_ms_anom_la}
\alias{pf_ms_anom_la}
\title{\strong{Multi-Site, Anomaly, Longitudinal}}
\usage{
pf_ms_anom_la(
  process_output,
  domain_filter,
  visit_filter,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{output of \code{pf_process} function}

\item{domain_filter}{one of the user provided domains in the process_output table to be used
to filter down the output}

\item{visit_filter}{one of the user provided visit types in the process_output table to be used
to filter down the output}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
returns three graphs:
\enumerate{
\item line graph that shows the smoothed proportion of a
domain across time computation with the Euclidean distance associated with each line
\item line graph that shows the raw proportion of a
domain across time computation with the Euclidean distance associated with each line
\item a bar graph with the Euclidean distance value for each site, with the average
proportion as the fill
}
}
\description{
\strong{Multi-Site, Anomaly, Longitudinal}
}
