% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output_subfuncs.R
\name{pf_ms_anom_nt}
\alias{pf_ms_anom_nt}
\title{\strong{Multi-Site, Anomaly Detection, No Time}}
\usage{
pf_ms_anom_nt(data_tbl, facet = NULL, visit_filter = "inpatient")
}
\arguments{
\item{data_tbl}{output from \code{pf_process} function}

\item{facet}{variables to facet (e.g., \code{domain}); vector of strings}

\item{visit_filter}{the single visit_type of interest to be used in the analysis}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the proportion of patients
for a given domain, and the size of the dot represents the mean proportion
across all sites

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if no groups were eligible for anomaly detection analysis, a heatmap summarizing
    the proportion of patients for a given domain & site and a dot plot summarizing
    the average standard deviation for each site are returned as an alternative
}\if{html}{\out{</div>}}
}
\description{
\strong{Multi-Site, Anomaly Detection, No Time}
}
