% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output_subfuncs.R
\name{pf_ms_exp_cs}
\alias{pf_ms_exp_cs}
\title{\strong{Multi-Site, Exploratory, Cross-Sectional}}
\usage{
pf_ms_exp_cs(data_tbl, output, facet, large_n = FALSE, large_n_sites = NULL)
}
\arguments{
\item{data_tbl}{output from \code{pf_process} function}

\item{output}{desired output - have 2 options:
\itemize{
\item \code{median_site_with0s}: specific site median, including patients with no evidence of patient fact
(e.g., if domain = labs, includes in the median all patients with and without any labs)
\item \code{median_site_without0s}: specific site median, not including patients without evidence of patient fact
(e.g., if domain = labs, only includes median for patients with evidence of a lab)
}}

\item{facet}{variables to facet (e.g., \code{domain}); vector of strings}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
a dot plot displaying the median facts per follow up for each domain
and site compared to the all-site median (star icon)
}
\description{
\strong{Multi-Site, Exploratory, Cross-Sectional}
}
