% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output_subfuncs.R
\name{pf_ms_exp_la}
\alias{pf_ms_exp_la}
\title{\strong{Multi-Site, Exploratory, Longitudinal}}
\usage{
pf_ms_exp_la(
  data_tbl,
  output,
  facet = NULL,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{data_tbl}{output from \code{pf_process} function}

\item{output}{desired output - have 2 options:
\itemize{
\item \code{median_fact_ct}: the median number of facts for each domain during the
specified time period
\item \code{sum_fact_ct}: the sum of facts for each domain during the specified time period
}}

\item{facet}{variables to facet (e.g., \code{domain}); vector of strings}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
line graph representing the output variable of interest across time
for each of the sites of interest; each site is represented by one line
}
\description{
\strong{Multi-Site, Exploratory, Longitudinal}
}
