% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output_subfuncs.R
\name{pf_ms_exp_nt}
\alias{pf_ms_exp_nt}
\title{#' \strong{Multi-Site, Exploratory, No Time}}
\usage{
pf_ms_exp_nt(data_tbl, output, facet)
}
\arguments{
\item{data_tbl}{output from \code{pf_process} function}

\item{output}{desired output - have 2 options:
\enumerate{
\item \code{median_site_with0s}: specific site median, including patients with no evidence of patient fact
(e.g., if domain = labs, includes in the median all patients with and without any labs)
\item \code{median_site_without0s}: specific site median, not including patients without evidence of patient fact
(e.g., if domain = labs, only includes median for patients with evidence of a lab)
}}

\item{facet}{variables to facet (e.g., \code{domain}); vector of strings}
}
\value{
a dot plot displaying the median facts per follow up for each domain
compared to the all-site median (star icon)
}
\description{
This chart will produce output for each site & domain. The data frame
that is listed as a parameter of the function should minimally contain a single
output for each domain and site combination. User can facet by visit type, age category,
or other stratifications. Multiple graphs can also be produced with varying facets.
}
