% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output.R
\name{pf_output}
\alias{pf_output}
\title{Patient Facts -- Output Generation}
\usage{
pf_output(
  process_output,
  output = NULL,
  date_breaks_str = "1 year",
  domain_filter = NULL,
  visit_filter = NULL,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{\emph{tabular input} || \strong{required}

The tabular output produced by \code{pf_process}

Note any patient-level results generated are not intended to be used with this function.}

\item{output}{\emph{string} || defaults to \code{NULL}

The name of the numerical variable from \code{process_output} that should be used to
generate the plot. This input is required for the following checks:
\itemize{
\item \verb{Single Site, Exploratory, Cross-Sectional}
\item \verb{Multi-Site, Exploratory, Cross-Sectional}
\item \verb{Single Site, Anomaly Detection, Cross-Sectional}
\item \verb{Single Site, Exploratory Longitudinal}
\item \verb{Multi Site Exploratory Longitudinal}
}}

\item{date_breaks_str}{\emph{string} || defaults to \verb{1 year}

A string that controls the time period division on the x-axis ('1 year', '3 months', etc).
This parameter is only required for the \verb{Single Site, Exploratory, Longitudinal} check.}

\item{domain_filter}{\emph{string} || defaults to \code{NULL}

A string indicating the domain of interest for plotting. This parameter
is required for the following checks:
\itemize{
\item \verb{Single Site, Anomaly Detection, Longitudinal}
\item \verb{Multi-Site, Anomaly Detection, Longitudinal}
}}

\item{visit_filter}{\emph{string} || defaults to \code{NULL}

A string indicating the visit type of interest for plotting. This parameter
is required for the following checks:
\itemize{
\item \verb{Multi-Site, Anomaly Detection, Cross-Sectional}
\item \verb{Single Site, Anomaly Detection, Longitudinal}
\item \verb{Multi-Site, Anomaly Detection, Longitudinal}
}}

\item{large_n}{\emph{boolean} || defaults to \code{FALSE}

For Multi-Site analyses, a boolean indicating whether the large N
visualization, intended for a high volume of sites, should be used. This
visualization will produce high level summaries across all sites, with an
option to add specific site comparators via the \code{large_n_sites} parameter.}

\item{large_n_sites}{\emph{vector} || defaults to \code{NULL}

When \code{large_n = TRUE}, a vector of site names that can add site-level information
to the plot for comparison across the high level summary information.}
}
\value{
This function will produce a graph to visualize the results
from \code{pf_process} based on the parameters provided. The default
output is typically a static ggplot or gt object, but interactive
elements can be activated by passing the plot through \code{make_interactive_squba}.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
Using the tabular output generated by \code{pf_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'patientfacts'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'pf_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

## turn of SQL trace for example
config('db_trace', FALSE)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                                      # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Execute `pf_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
pf_process_example <- pf_process(cohort = cohort,
                                 study_name = 'example_study',
                                 multi_or_single_site = 'single',
                                 anomaly_or_exploratory = 'exploratory',
                                 visit_type_table =
                                   patientfacts::pf_visit_file_omop,
                                 omop_or_pcornet = 'omop',
                                 visit_types = c('all'),
                                 domain_tbl = patientfacts::pf_domain_file \%>\%
                                   dplyr::filter(domain == 'diagnoses')) \%>\%
  suppressMessages()

pf_process_example

#' Execute `pf_output` function
#' The output was edited for a better indication of what the visualization will
#' look like.
#' The 0s are a limitation of the small sample data set used for this example
pf_output_example <- pf_output(process_output = pf_process_example \%>\%
                                 dplyr::mutate(median_site_without0s = 4),
                                 ## tweak synthetic output for example
                               output = 'median_site_without0s')

pf_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(pf_output_example)
}
