% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_process.R
\name{pf_process}
\alias{pf_process}
\title{Patient Facts}
\usage{
pf_process(
  cohort = cohort,
  study_name = "my_study",
  omop_or_pcornet = "omop",
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  time = FALSE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year",
  p_value = 0.9,
  age_groups = NULL,
  patient_level_tbl = FALSE,
  visit_types = c("outpatient", "inpatient"),
  domain_tbl = patientfacts::pf_domain_file,
  visit_tbl = cdm_tbl("visit_occurrence"),
  visit_type_table = patientfacts::pf_visit_file_omop
)
}
\arguments{
\item{cohort}{\emph{tabular input} || \strong{required}

The cohort to be used for data quality testing. This table should contain,
at minimum:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}

Note that the start and end dates included in this table will be used to
limit the search window for the analyses in this module.}

\item{study_name}{\emph{string} || defaults to \code{my_study}

A string identifier for the name of your study}

\item{omop_or_pcornet}{\emph{string} || \strong{required}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data
\itemize{
\item \code{omop}: run the \code{\link[=pf_process_omop]{pf_process_omop()}} function against an OMOP CDM instance
\item \code{pcornet}: run the \code{\link[=pf_process_pcornet]{pf_process_pcornet()}} function against a PCORnet CDM instance
}}

\item{multi_or_single_site}{\emph{string} || defaults to \code{single}

A string, either \code{single} or \code{multi}, indicating whether a single-site or
multi-site analysis should be executed}

\item{anomaly_or_exploratory}{\emph{string} || defaults to \code{exploratory}

A string, either \code{anomaly} or \code{exploratory}, indicating what type of results
should be produced.

Exploratory analyses give a high level summary of the data to examine the
fact representation within the cohort. Anomaly detection analyses are
specialized to identify outliers within the cohort.}

\item{time}{\emph{boolean} || defaults to \code{FALSE}

A boolean to indicate whether to execute a longitudinal analysis}

\item{time_span}{\emph{vector - length 2} || defaults to \code{c('2012-01-01', '2020-01-01')}

A vector indicating the lower and upper bounds of the time series for longitudinal analyses}

\item{time_period}{\emph{string} || defaults to \code{year}

A string indicating the distance between dates within the specified time_span.
Defaults to \code{year}, but other time periods such as \code{month} or \code{week} are
also acceptable}

\item{p_value}{\emph{numeric} || defaults to \code{0.9}

The p value to be used as a threshold in the Multi-Site,
Anomaly Detection, Cross-Sectional analysis}

\item{age_groups}{\emph{tabular input} || defaults to \code{NULL}

If you would like to stratify the results by age group, create a table or
CSV file with the following columns and use it as input to this parameter:
\itemize{
\item \code{min_age} | \emph{integer} | the minimum age for the group (i.e. 10)
\item \code{max_age} | \emph{integer} | the maximum age for the group (i.e. 20)
\item \code{group} | \emph{character} | a string label for the group (i.e. 10-20, Young Adult, etc.)
}

If you would \emph{not} like to stratify by age group, leave as \code{NULL}}

\item{patient_level_tbl}{\emph{boolean} || defaults to \code{FALSE}

A boolean indicating whether an additional table with patient level results should be output.

If \code{TRUE}, the output of this function will be a list containing both the summary and patient level
output. Otherwise, this function will just output the summary dataframe}

\item{visit_types}{\emph{string or vector} || defaults to \code{c('outpatient', 'inpatient')}

A string or vector of visit types by which the output should be stratified. Each
visit type listed in this parameter should match an associated visit type defined
in the \code{visit_type_table}}

\item{domain_tbl}{\emph{tabular input} || \strong{required}

A table that defines the fact domains to be investigated in the analysis. This
input should contain:
\itemize{
\item \code{domain} | \emph{character} | a string label for the domain being examined (i.e. prescription drugs)
\item \code{domain_tbl} | \emph{character} | the CDM table where information for this domain can be found (i.e. drug_exposure)
\item \code{filter_logic} | \emph{character} | logic to be applied to the domain_tbl in order to achieve the definition of interest; should be written as if you were applying it in a dplyr::filter command in R
}}

\item{visit_tbl}{\emph{tabular input} || defaults to \code{cdm_tbl('visit_occurrence')}

The CDM table with visit information (i.e. visit_occurrence or encounter)}

\item{visit_type_table}{\emph{tabular input} || \strong{required}

A table that defines visit types of interest called in \code{visit_types.} This input
should contain:
\itemize{
\item \code{visit_concept_id} / \code{visit_detail_concept_id} or \code{enc_type} | \emph{integer} or \emph{character} | the \verb{visit_(detail)_concept_id} or \code{enc_type} that represents the visit type of interest (i.e. 9201 or IP)
\item \code{visit_type} | \emph{character} | the string label to describe the visit type
}}
}
\value{
This function will return a dataframe summarizing the
distribution of facts per visit type for each user defined variable.
It can also optionally return un-summarized patient-level
distributions. For a more detailed description of output specific to
each check type, see the PEDSpace metadata repository
}
\description{
This is a completeness module that will compute the number of facts per years of follow-up for each patient in
a cohort. The user will provide the domains (\code{domain_tbl}) and visit types (\code{visit_type_tbl}) of interest.
Sample versions of these inputs are included as data in the package and are accessible with \verb{patientfacts::}.
Results can optionally be stratified by site, age group, visit type, and/or time. This function is compatible with
both the OMOP and the PCORnet CDMs based on the user's selection.
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'patientfacts'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'pf_process_test',
                      working_directory = getwd(),
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = 'extdata',
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                                      # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Execute `pf_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
pf_process_example <- pf_process(cohort = cohort,
                                 study_name = 'example_study',
                                 multi_or_single_site = 'single',
                                 anomaly_or_exploratory = 'exploratory',
                                 visit_type_table =
                                   patientfacts::pf_visit_file_omop,
                                 omop_or_pcornet = 'omop',
                                 visit_types = c('all'),
                                 domain_tbl = patientfacts::pf_domain_file \%>\%
                                   dplyr::filter(domain == 'diagnoses'))

pf_process_example

#' Execute `pf_output` function
#' The output was edited for a better indication of what the visualization will
#' look like.
#' The 0s are a limitation of the small sample data set used for this example
pf_output_example <- pf_output(process_output = pf_process_example \%>\%
                                 dplyr::mutate(median_site_without0s = 4),
                               output = 'median_site_without0s')

pf_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(pf_output_example)
}
