% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output_subfuncs.R
\name{pf_ss_anom_at}
\alias{pf_ss_anom_at}
\title{\strong{Single Site, Anomaly, Over Time}}
\usage{
pf_ss_anom_at(data_tbl, facet, visit_filter, domain_filter)
}
\arguments{
\item{data_tbl}{output from \code{pf_process} function}

\item{facet}{variables to facet (e.g., \code{domain}); vector of strings}

\item{visit_filter}{the single visit type of focus for the output}

\item{domain_filter}{the single domain of focus for the output}
}
\value{
if analysis was executed by year or greater, a P Prime control chart
is returned with outliers marked with orange dots

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if analysis was executed by month or smaller, an STL regression is
    conducted and outliers are marked with red dots. the graphs representing
    the data removed in the regression are also returned
}\if{html}{\out{</div>}}
}
\description{
\strong{Single Site, Anomaly, Over Time}
}
