% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output_subfuncs.R
\name{pf_ss_anom_cs}
\alias{pf_ss_anom_cs}
\title{\strong{Single Site, Anomaly Detection, Cross-Sectional}}
\usage{
pf_ss_anom_cs(data_tbl, output, facet = c("visit_type"))
}
\arguments{
\item{data_tbl}{output from \code{pf_process} function}

\item{output}{desired output - have 4 options:
\itemize{
\item \code{outlier_fact}: the number of facts overall (i.e. not grouped by site) that fall 2 SD
away from the mean
\item \code{prop_outlier_fact}: the proportion of facts overall that fall 2 SD away from the mean
\item \code{outlier_site_fact}: the number of facts per site that fall 2 SD away from the mean
\item \code{prop_outlier_site_fact}: the proportion of facts per site that fall 2 SD away from the mean
}}

\item{facet}{variables to facet (e.g., \code{domain}); vector of strings}
}
\value{
a bar graph displaying the output value of interest, which represents
patients falling +/- 2 standard deviations away from the mean facts
per follow-up for a given domain
}
\description{
\strong{Single Site, Anomaly Detection, Cross-Sectional}
}
\keyword{internal}
