% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output_subfuncs.R
\name{pf_ss_exp_nt}
\alias{pf_ss_exp_nt}
\title{\strong{Single-Site, Exploratory, No Time}}
\usage{
pf_ss_exp_nt(data_tbl, output, facet = c("domain"))
}
\arguments{
\item{data_tbl}{output from previous function;
requires input for one site; can create multiple graphs
for different grouped variables}

\item{output}{desired output - have 2 options:
\enumerate{
\item \code{median_site_with0s}: specific site median, including patients with no evidence of patient fact
(e.g., if domain = labs, includes in the median all patients with and without any labs)
\item \code{median_site_without0s}: specific site median, not including patients without evidence of patient fact
(e.g., if domain = labs, only includes median for patients with evidence of a lab)
}}

\item{facet}{variables to facet (e.g., \code{domain}); vector of strings}
}
\value{
a bar graph displaying the median facts per follow-up for each domain
and visit_type
}
\description{
This chart will produce output for each domain. The data frame
that is listed as a parameter of the function should minimally contain a single
output for each domain. User can facet by site, age category,
or other stratifications. Multiple graphs can also be produced (e.g.,
one graph faceting by site, or creaeting separate output for each site)
}
