% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pf_visit_file_pcornet}
\alias{pf_visit_file_pcornet}
\title{Patient Facts Visit File Sample -- PCORnet version}
\format{
\subsection{pf_visit_file_pcornet}{

\describe{
\item{enc_type}{The enc_type as it appears in the encounter table}
\item{visit_type}{A string to label the visit type of the enc_type; this string is what should be referenced in the \code{visit_types} parameter}
}
}
}
\usage{
pf_visit_file_pcornet
}
\description{
A sample version of the file structure expected for the \code{visit_type_tbl}
parameter in the \code{pf_process} function where the selected CDM is PCORnet.
The user should recreate this file and include their own domain definitions.
}
\keyword{datasets}
