% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_process_subfuncs.R
\name{combine_study_facts}
\alias{combine_study_facts}
\title{Combine the output of compute_pf or compute_pf_for_fot into one dataframe}
\usage{
combine_study_facts(pf_tbl, study_abbr, domain_list, time, visit_type_list)
}
\arguments{
\item{pf_tbl}{the table output by \code{compute_pf} or \code{compute_pf_for_fot}}

\item{study_abbr}{string that corresponds with the study/studies of interest}

\item{domain_list}{a table with definitions for each domain (i.e. patientfacts::pf_domain_file)}

\item{time}{a logical indicating whether the analysis was conducted over time}

\item{visit_type_list}{the types of visits in the PF output; defaults to:
inpatient, outpatient, other_visit, all}
}
\value{
a dataframe containing information associated with all sites, domains,
and visit types for the study listed in \code{study_abbr}
}
\description{
Combine the output of compute_pf or compute_pf_for_fot into one dataframe
}
