% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_process_subfuncs.R
\name{compute_pf_medians}
\alias{compute_pf_medians}
\title{Compute median fact counts by visit & domain and by site, visit, & domain}
\usage{
compute_pf_medians(data_input, site_col, agegrp = NULL)
}
\arguments{
\item{data_input}{input table generated by \code{loop_through_visits} and reduced with \code{combine_study_facts}}

\item{site_col}{the column in the data that holds site names}

\item{agegrp}{If you would like to stratify the results by age group, create a table or CSV file with the following
columns and include it as the \code{agegrp} function parameter:
\itemize{
\item \code{min_age}: the minimum age for the group (i.e. 10)
\item \code{max_age}: the maximum age for the group (i.e. 20)
\item \code{group}: a string label for the group (i.e. "10-20", "Young Adult", etc.)
}}
}
\value{
dataframe that contains the total median number of facts for that visit & domain
as well as the median number of facts for the site, visit, & domain for each
study. The medians are compute both with patients who do not have evidence of the domain
(with 0s) and without those patients (without 0s)
}
\description{
Compute median fact counts by visit & domain and by site, visit, & domain
}
