% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output.R
\name{pf_output}
\alias{pf_output}
\title{Clinical Facet per Patient -- Output Generation}
\usage{
pf_output(
  process_output,
  output_function,
  output,
  facet = NULL,
  time_span = c("2012-01-01", "2023-01-01"),
  date_breaks_str = "1 year",
  domain_filter = "conditions_all",
  visit_filter = "outpatient"
)
}
\arguments{
\item{process_output}{\itemize{
\item the summary dataframe output by the \code{pf_process} function.

\if{html}{\out{<div class="sourceCode">}}\preformatted{              Note any intermediate table generated is not intended to be used with this function.
}\if{html}{\out{</div>}}
}}

\item{output_function}{\itemize{
\item the name of the output function that should be used provided in the \code{parameter_summary} csv
file that is output to the provided results folder after running the \code{pf_process} function
}}

\item{output}{\itemize{
\item which output variable you would like to use in the graphs. available options based on check
configuration are provided in the \code{parameter_summary} csv file
}}

\item{facet}{\itemize{
\item the variables by which you would like to facet the graph. available and/or recommended options for
faceting variables are provided in the \code{parameter_summary} csv file
}}

\item{time_span}{\itemize{
\item the length of time that should be displayed for relevant over-time graphs. this time period should
either be the same as or a subset of the time period used in the \code{pf_process} function
}}

\item{date_breaks_str}{\itemize{
\item only for single-site, exploratory, over time; a string that informs the program how the
time period should be divided (i.e. '1 year', '3 months', etc). Defaults to 1 year.
}}

\item{domain_filter}{for \code{pf_ms_anom_at} only, the single domain to which the graph should be filtered for displaying
AUC values}

\item{visit_filter}{for \code{pf_ms_anom_at} only, the single visit type to which the graph should be filtered for displaying
AUC values}
}
\value{
"raw" output that can be called within the R environment to generate the graph in the Viewer window
}
\description{
Clinical Facet per Patient -- Output Generation
}
