% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output.R
\name{pf_output}
\alias{pf_output}
\title{Clinical Facet per Patient -- Output Generation}
\usage{
pf_output(
  process_output,
  output_function,
  output,
  date_breaks_str = "1 year",
  domain_filter = "conditions_all",
  visit_filter = "outpatient"
)
}
\arguments{
\item{process_output}{the summary dataframe output by the \code{pf_process} function.

\if{html}{\out{<div class="sourceCode">}}\preformatted{                  Note any patient-level results generated are not intended to be used with this function.
}\if{html}{\out{</div>}}}

\item{output_function}{\itemize{
\item the name of the output function that should be executed, provided in the message output
to the console after \code{pf_process} has been executed
}}

\item{output}{the numerical variable in \code{process_output} that should be used to generate the graph}

\item{date_breaks_str}{for \code{pf_ss_exp_at} only, a string that informs the program how the
time period should be divided (i.e. '1 year', '3 months', etc). Defaults to 1 year.}

\item{domain_filter}{for \code{pf_ms_anom_at} only, the single domain to which the graph should be filtered for displaying
Euclidean distance values}

\item{visit_filter}{for \code{pf_ms_anom_at} only, the single visit type to which the graph should be filtered for displaying
Euclidean distance values}
}
\value{
a graph summarizing the data output by \code{pf_process}; see individual output functions for specific details
}
\description{
Clinical Facet per Patient -- Output Generation
}
