% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_output.R
\name{pf_output}
\alias{pf_output}
\title{Clinical Facts per Patient -- Output Generation}
\usage{
pf_output(
  process_output,
  output_function,
  output,
  date_breaks_str = "1 year",
  domain_filter = NULL,
  visit_filter = NULL
)
}
\arguments{
\item{process_output}{the summary dataframe output by the \code{pf_process} function.

Note any patient-level results generated are not intended to be used with this function.}

\item{output_function}{the name of the output function that should be executed, provided in the message output
to the console after \code{pf_process} has been executed}

\item{output}{the numerical variable in \code{process_output} that should be used to generate the graph}

\item{date_breaks_str}{for \code{pf_ss_exp_la} only, a string that informs the program how the
time period should be divided (i.e. '1 year', '3 months', etc). Defaults to 1 year.}

\item{domain_filter}{for \code{pf_ms_anom_la} only, the single domain to which the graph should be filtered for displaying
Euclidean distance values}

\item{visit_filter}{for \code{pf_ms_anom_la} only, the single visit type to which the graph should be filtered for displaying
Euclidean distance values}
}
\value{
a graph summarizing the data output by \code{pf_process}; see individual output functions for specific details
}
\description{
Using the tabular output generated by \code{pf_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_ssdqa()}
}
