% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_process.R
\name{pf_process}
\alias{pf_process}
\title{Clinical Facts per Patient}
\usage{
pf_process(
  cohort = cohort,
  study_name = "glom",
  patient_level_tbl = FALSE,
  visit_types = c("outpatient", "inpatient"),
  multi_or_single_site = "multi",
  time = FALSE,
  time_span = c("2014-01-01", "2023-01-01"),
  time_period = "year",
  p_value = 0.9,
  age_groups = NULL,
  anomaly_or_exploratory = "anomaly",
  domain_tbl = patientfacts::pf_domain_file,
  visit_type_table = read_codeset("pf_visit_types", "ic")
)
}
\arguments{
\item{cohort}{\itemize{
\item A dataframe with the cohort of patients for your study. Should include the columns:
- @person_id
- @start_date
- @end_date
- @site
}}

\item{study_name}{\itemize{
\item A custom string label with the name of your study
}}

\item{patient_level_tbl}{\itemize{
\item logical indicating whether an additional table with patient level results should be output;
if TRUE, an R dataframe called \code{pf_patient_level_results} will appear in the global environment
in addition to the standard summary output
}}

\item{visit_types}{\itemize{
\item A list of visit types by which the output should be stratified. Options for visit types can be found or adjusted
in the provided \code{pf_visit_types.csv} file. If you would not like to stratify your results by visit type, please
set the visit_types argument equal to \code{all}
}}

\item{multi_or_single_site}{\itemize{
\item Option to run the function on a single vs multiple sites
- @single - run the function for a single site
- @multi - run the function for multiple sites
}}

\item{time}{\itemize{
\item a logical that tells the function whether you would like to look at the output over time
}}

\item{time_span}{\itemize{
\item when time = TRUE, this argument defines the start and end dates for the time period of interest. should be
formatted as c(start date, end date) in yyyy-mm-dd date format.
}}

\item{time_period}{\itemize{
\item when time = TRUE, this argument defines the distance between dates within the specified time period. defaults
to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable
}}

\item{p_value}{the p value to be used as a threshold in the multi-site anomaly detection analysis}

\item{age_groups}{\itemize{
\item If you would like to stratify the results by age group, fill out the provided \code{age_group_definitions.csv} file
with the following information:
- @min_age: the minimum age for the group (i.e. 10)
- @max_age: the maximum age for the group (i.e. 20)
- @group: a string label for the group (i.e. 10-20, Young Adult, etc.)

\if{html}{\out{<div class="sourceCode">}}\preformatted{              Then supply this csv file as the age_groups argument (i.e. read.csv('path/to/age_group_definitions.csv'))

              If you would not like to stratify by age group, leave the argument as NULL
}\if{html}{\out{</div>}}
}}

\item{anomaly_or_exploratory}{\itemize{
\item Option to conduct an exploratory or anomaly detection analysis. Exploratory analyses give a high
level summary of the data to examine the fact representation within the cohort. Anomaly detection
analyses are specialized to identify outliers within the cohort.
}}

\item{domain_tbl}{\itemize{
\item a csv file that defines the domains where facts should be identified. defaults to the provided
\code{pf_domains.csv} file, which contains the following fields:
- @domain: a string label for the domain being examined (i.e. prescription drugs)
- @default_tbl: the CDM table where information for this domain can be found (i.e. drug_exposure)
- @field_name: if filtering is required, which field should be used to filter the default_tbl (i.e. drug_type_concept_id)
- @field_filter: if filtering is required, a string with codes that will filter the default_tbl to the
domain of interest (i.e. "38000177")

\if{html}{\out{<div class="sourceCode">}}\preformatted{             This CSV can be altered to fit the users needs, or another csv with the same columns and formatting can be supplied.
}\if{html}{\out{</div>}}
}}

\item{visit_type_table}{\itemize{
\item a csv file that defines available visit types that are called in @visit_types. defaults to the provided
\code{pf_visit_types} file, which contains the following fields:
- @visit_concept_id: the visit_concept_id that represents the visit type of interest (i.e. 9201)
- @visit_type: the string label to describe the visit type; this label can be used multiple times
within the file if multiple visit_concept_ids represent the visit type

\if{html}{\out{<div class="sourceCode">}}\preformatted{                    This CSV can be altered to fit the users needs, or another csv with the same columns and formatting can be supplied.
}\if{html}{\out{</div>}}
}}
}
\value{
if intermediate_tbl is not NULL, an intermediate results table is returned in the user-defined format

a dataframe with summary results (i.e. medians) that can be used as the input for \code{pf_output_gen} to generate graphical output

a CSV file with a summary of the parameters used to configure the function and recommendations for how to configure the
\code{pf_output_gen} function parameters; written to the provided results folder
}
\description{
This is a completeness check that will compute the number of facts per years of follow-up for each patient in
a cohort. The user will provide the domains of interest in a provided CSV file. Results can optionally be stratified
by site, age group, visit type, codeset utilization, and/or time.
}
